chore(plugins/plugin-client-common): Clean out old global sidecar code
Fixes #6077
