+++ plugins/plugin-client-common/web/css/static/Form.scss
@@ -1,4 +1,4 @@
-[kui-theme-style] {
+body[kui-theme-style] {
   .kui--form-section {
     &:not(:first-child) {
       padding-top: 1em;
