+++ plugins/plugin-client-common/web/css/static/Tooltip.scss
@@ -1,6 +1,6 @@
 @import 'carbon-components/scss/components/tooltip/_tooltip.scss';
 
-[kui-theme-style] {
+body[kui-theme-style] {
   .bx--tooltip__trigger {
     &.bx--tooltip--bottom.bx--tooltip--align-end {
       .bx--assistive-text {
