+++ plugins/plugin-client-common/web/scss/components/CodeSnippet/PatternFly.scss
@@ -16,11 +16,11 @@
 
 @import '../../PatternFly/common';
 
-[kui-theme-style] .pf-c-clipboard-copy__expandable-content {
+body[kui-theme-style] .pf-c-clipboard-copy__expandable-content {
   color: var(--pf-global--Color--100);
 }
 
-[kui-theme-style='dark'] .pf-c-clipboard-copy__expandable-content {
+body[kui-theme-style='dark'] .pf-c-clipboard-copy__expandable-content {
   color: var(--pf-global--Color--300);
 }
