+++ plugins/plugin-client-common/web/scss/components/DropDown/Carbon.scss
@@ -22,7 +22,7 @@ $bg-open-color: var(--color-dropdown);
 $button-hover-color: var(--color-table-border3);
 $option-hover-color: var(--color-base02);
 
-[kui-theme-style] {
+body[kui-theme-style] {
   .sidecar-bottom-stripe-button .bx--overflow-menu.bx--overflow-menu--open svg,
   .sidecar-bottom-stripe-button:hover .bx--overflow-menu.bx--overflow-menu--open svg,
   .kui--toolbar-button-with-icon:hover .bx--overflow-menu.bx--overflow-menu--open svg {
