+++ plugins/plugin-client-common/web/scss/components/List/Carbon.scss
@@ -16,7 +16,7 @@
 
 @import '~carbon-components/scss/components/list/_list.scss';
 
-[kui-theme-style] {
+body[kui-theme-style] {
   .bx--list--ordered,
   .bx--list--unordered {
   }
