+++ plugins/plugin-client-common/web/scss/components/Table/common.scss
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-[kui-theme-style] .kui--table-like {
+body[kui-theme-style] .kui--table-like {
   font-family: var(--font-sans-serif);
 }
